Mid-Working Secretary Selector For Add
So i tried fixing the secretary error which was it was not listing all the secretaries that existed in the database, for now, it shows a list, mind you, not a pretty one to say the list, its not formatted, maybe its because im using threaded for the search? Anyway, its in progress for now, ill work on it more later on, i will be working on the CA for Adv Net now, but i will be getting back shortly. Cheers.